requirements.txt
alembic==1.13.2
Babel==2.15.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
cloudinary==1.40.0
cryptography==42.0.8
Flask==3.0.3
Flask-Admin==1.6.1
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.1
setuptools==70.2.0
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
